requirements.txt
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
underthesea==6.8.4
underthesea_core==1.0.4
urllib3==2.5.0
vietnamese==0.0.8
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
flask==3.1.1
flask_cors==5.0.0
nudenet==3.4.2
sentence_splitter==1.4
pillow==11.3.0